School-Graduate_thesis
The project is a one-page site on the topic of "Špela beauty" and was made for the needs of a beauty salon. The web application was developed in OOPHP using MySQL database and by using MVC pattern in Laravel framework. The application is divided into several segments, which are divided according to user roles and thus according to certain authorizations. The goals of the web application are the presentation of the beauty salon and management of the site content, as well as the possibility of giving basic advice to users through a blog and the possibility of scheduling treatments. The segment of application for the unauthorized users includes a presentation and description of available treatments in the salon, a location realized by using Google geolocation, as well as access to the pages of the beauty salon on social networks. Also, in this segment, the user is able to review and search the blog, as well as to schedule a treatment in the salon in one of the available appointments, which was realized through Ajax. The segment for bloggers includes viewing and searching the blog as an unauthorized user can view it. On the blogger page, it is possible to enter a new blog, change the existing one, search all blogs by the date, which was realized in Ajax but also to deactivate or delete the existing blog. The segment related to the employee allows scheduling treatment in one of the available appointments realized by Ajax, review of scheduled appointments and their search by date, as well as a display of canceled appointments and their search by date. Scheduled appointments can be changed and canceled here. Searches on the employee page were realized through Ajax. The segment for administrator is the admin panel where you can manage the content of the site. The content that can be changed concerns the possibility of viewing, creating, modifying, and deactivating or deleting the type of treatment to which a particular treatment belongs, the treatment itself, users, slider images, blogs, scheduled, and canceled appointments, and beauty salon information. On the administration panel, it is possible to do a search by date for content such as the blog as well as scheduled and canceled appointments, which were realized through Ajax. Checking the data entered through forms on the application are realized on the server side with the help of regular expressions.
Technologies - HTML, CSS, Bootstrap, PHP, MVC, Laravel framework, JavaScript, JQuery, Ajax, Blade, MySql